old-tests/submission/W3C/domhtml/HTMLImageElement12.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-img-element.html#dom-img-width" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img src="/support/images/yellow-rectangle.png" width="115">
</p>



<script>
var domtest = async_test("The width attribute overrides the natural \"width\" of the image.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vwidth;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vwidth = testNode.width;

      assert_equals(vwidth,115);
       
    domtest.done();
}));

</script>
<p>The width attribute overrides the natural "width" of the image.  

    Retrieve the width attribute and examine it's value.</p>
<div id="log"></div>
</body></html>
